Initial patch to fix trigger spam
Currently if the framerate is still above or below the highest/lowest value after triggering, it will keep on triggering the same commands and messages every time. This patch is intended to fix this by taking into account what the last profile was and to cancel running the command and showing the chat message if already ran and shown last time.